CASMCMS-7783: Patch BSS metadata key for CFS Instead of Put
This mod changes the population behavior for cfs-trust; this migrates
the behavior over to a patch operation in order to be a good citizen for
any service that ends up populating information afterwards. Normally, we
don't re-PUT once our key is there, and since our services is one of the
first to push content into the bss metadata service, there is no
problem.

However, during an upgrade or scenarios where cfs-trust metadata goes
missing, subsequent PUT operations zero any incidental keys stored in
BSS metadata global section; this information then needs to be recovered
(sometimes manually).

Finally, this mod updates the behavior of cfs-trust and changes it to
re-populate our missing key if it ever goes missing for whatever reason.
Previously, a pod restart would accomplish this similar behavior but now
that is no longer needed. The existing resigning/republishing loop is
used to accomplish this, so BSS metadata healing happens every 6 hours
or so.