CASMINST-5443:
Ensure that the csm-python-module can be referenced when we building
python packages that are not part of building Docker images.
We need to have the .netrc file in the environment, which is a Docker
container, where we are building the python package.
Note the distinction, we are building the python package inside a
Docker container, but not as a part of building the Docker image.